requirements.txt
# Core dependencies
apache-airflow==2.7.3
apache-airflow-providers-mongo==3.4.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3
isodate==0.6.1

# Data quality
soda-core-mongo==3.0.45
soda-core-scientific==3.0.45

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
unittest-xml-reporting==3.2.0
mongomock==4.1.2

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0

# Logging
structlog==23.2.0
colorlog==6.8.0

# Utilities
click==8.1.7
pyyaml==6.0.1
jsonschema==4.20.0
